Expose Session Manager, associated getter/setter hook, and associated types externally, and allow Session Manager to be used independently/earlier in IOPA middleware chain
