+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iopa-bot",
-  "version": "2.1.13",
+  "version": "2.2.2",
   "main": "dist/index.js",
   "ts:main": "src/index.ts",
   "types": "src/index.ts",
